requirements.txt
kafka-python
pyspark==3.5.1
delta-spark